requirements.txt
streamlit>=1.32.0
requests>=2.31.0
anthropic>=0.25.0
openai>=1.14.0
python-dotenv>=1.0.0
Pillow>=10.0.0
